# openapi.action.yaml
openapi: 3.1.0
info:
  title: CyLaw / CyLII Legal Research
  description: Read-only Cyprus legal research over public CyLaw and CyLII material, including full-text case-law search and New Civil Procedure Rules retrieval. Not affiliated with the Cyprus Bar Association.
  version: 0.3.0
servers:
  - url: https://cylaw-api.onrender.com
paths:
  /search/cylaw:
    get:
      operationId: searchCylawFullText
      summary: Search the full text of Cyprus judgments on CyLaw
      parameters:
        - {name: query, in: query, required: true, schema: {type: string}, description: "Use Greek legal terms where possible, e.g. απόφαση ερήμην, σημείωμα εμφάνισης, παραμερισμός."}
        - {name: collection, in: query, required: false, schema: {type: string, default: supreme, enum: [supreme, courtOfAppeal, supremeAdministrative, administrativeCourtOfAppeal, apofaseis/aad]}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 10, minimum: 1, maximum: 20}}
      responses:
        '200': {description: Relevance-ranked case results, content: {application/json: {schema: {$ref: '#/components/schemas/CaseHitList'}}}}
  /search/new-cpr:
    get:
      operationId: searchNewCivilProcedureRules
      summary: Find relevant public New Civil Procedure Rules material
      parameters:
        - {name: query, in: query, required: true, schema: {type: string}, description: "Rule topic in Greek or English, for example default appearance, default judgment, set aside, service, or small claims."}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 10, minimum: 1, maximum: 20}}
      responses:
        '200': {description: Relevant rules or forms with text, content: {application/json: {schema: {$ref: '#/components/schemas/RuleHitList'}}}}
  /browse/cylii:
    get:
      operationId: browseCylii
      summary: List or filter cases from a CyLII section
      parameters:
        - {name: section, in: query, required: true, schema: {type: string}, description: "For example: cy/cases/diait/aap"}
        - {name: year, in: query, required: false, schema: {type: integer}}
        - {name: keyword, in: query, required: false, schema: {type: string}}
      responses:
        '200': {description: Matching cases, content: {application/json: {schema: {$ref: '#/components/schemas/CaseHitList'}}}}
  /browse/cylaw:
    get:
      operationId: browseCylaw
      summary: List or filter cases from a CyLaw static index page
      parameters:
        - {name: index_path, in: query, required: true, schema: {type: string}, description: "For example: supreme/index_2024.html"}
        - {name: keyword, in: query, required: false, schema: {type: string}}
      responses:
        '200': {description: Matching cases, content: {application/json: {schema: {$ref: '#/components/schemas/CaseHitList'}}}}
  /document:
    get:
      operationId: getDocument
      summary: Fetch the full text of a CyLaw or CyLII document
      parameters:
        - {name: url, in: query, required: true, schema: {type: string, format: uri}, description: "An HTTPS URL on cylaw.org or cylii.org"}
      responses:
        '200': {description: Document text, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
  /sections:
    get:
      operationId: listKnownSections
      summary: List starter CyLII section paths
      responses:
        '200':
          description: Known section paths
          content:
            application/json:
              schema:
                type: object
                properties:
                  supreme_court: {type: string}
                  administrative_court: {type: string}
                  tender_review_authority: {type: string}
                  competition_commission: {type: string}
components:
  schemas:
    CaseHit:
      type: object
      required: [number, title, url, source]
      properties:
        number: {type: string}
        title: {type: string}
        url: {type: string, format: uri}
        source: {type: string, enum: [cylaw, cylii]}
    CaseHitList:
      type: array
      items: {$ref: '#/components/schemas/CaseHit'}
    RuleHit:
      type: object
      properties:
        number: {type: string}
        title: {type: string}
        url: {type: string, format: uri}
        text: {type: string}
    RuleHitList:
      type: array
      items: {$ref: '#/components/schemas/RuleHit'}
    Document:
      type: object
      required: [url, title, text]
      properties:
        url: {type: string, format: uri}
        title: {type: string}
        text: {type: string}